tests/test_optimizer.py
import unittest
from optimizer import Optimizer

class TestOptimizer(unittest.TestCase):
    def setUp(self):
        self.optimizer = Optimizer()

    def test_optimize_inventory(self):
        # Test case 1
        result1 = self.optimizer.optimize_inventory(data1)
        self.assertEqual(result1, expected_optimization1)

        # Test case 2
        result2 = self.optimizer.optimize_inventory(data2)
        self.assertEqual(result2, expected_optimization2)

    def test_optimize_transportation(self):
        # Test case 1
        result1 = self.optimizer.optimize_transportation(data1)
        self.assertEqual(result1, expected_optimization1)

        # Test case 2
        result2 = self.optimizer.optimize_transportation(data2)
        self.assertEqual(result2, expected_optimization2)

    # Add more test methods for other optimizer functionalities

if __name__ == '__main__':
    unittest.main()